Add multiple changes

+++ src/pages/api/webhooks/email-engine.ts
@@ -1,27 +1,78 @@
-import { ActivityStatus } from '@prisma/client';
+import { AccountState, ActivityStatus } from '@prisma/client';
 import omit from 'lodash/omit';
 
 import { logger } from '@utils/logger';
 import { prisma } from '@server/db';
 import { WEBHOOKS } from '@utils/webhooks';
+import { updateAccountState } from '@webhooks/updateAccountState';
 
+import type { Prisma } from '@prisma/client';
 import type { NextApiRequest, NextApiResponse } from 'next';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   logger.info({ webhook: req.body }, `Webhook event`);
-  let payload;
 
   switch (req.body.event) {
-    case WEBHOOKS.ACCOUNT_DELETED: {
-      payload = await prisma.account.delete({ where: { id: req.body.account } });
-      logger.info({ webhook: req.body, payload }, `Successful webhook event ${WEBHOOKS.ACCOUNT_DELETED} executed`);
+    case WEBHOOKS.ACCOUNT_CONNECT_ERROR: {
+      const updatedAccount = await updateAccountState({
+        accountId: req.body.account,
+        state: AccountState.DISCONNECTED,
+      });
+      logger.info(
+        { webhook: req.body, updatedAccount },
+        `Webhook ${req.body.event}: Connection error to Email Engine service.`,
+      );
       break;
     }
-    case WEBHOOKS.TRACK_OPEN: {
+    case WEBHOOKS.ACCOUNT_AUTH_ERROR: {
+      const updatedAccount = await updateAccountState({ accountId: req.body.account, state: AccountState.ERROR });
+      logger.info(
+        { webhook: req.body, updatedAccount },
+        `Webhook ${req.body.event}: Connection auth error to Email Engine service.`,
+      );
+      break;
+    }
+    case WEBHOOKS.ACCOUNT_ADDED: {
+      const updatedAccount = await updateAccountState({ accountId: req.body.account, state: AccountState.CONNECTING });
+      logger.info(
+        { webhook: req.body, updatedAccount },
+        `Webhook ${req.body.event}: Connecting new account to Email Engine service.`,
+      );
+      break;
+    }
+    case WEBHOOKS.ACCOUNT_INITIALIZED:
+    case WEBHOOKS.ACCOUNT_AUTH_SUCCESS: {
+      const updatedAccount = await updateAccountState({ accountId: req.body.account, state: AccountState.CONNECTED });
+      logger.info({ webhook: req.body, updatedAccount }, `Webhook ${req.body.event}: Account successfully connected`);
+      break;
+    }
+    case WEBHOOKS.ACCOUNT_DELETED: {
+      const account = await prisma.account.findUnique({ where: { id: req.body.account } });
+      if (account) {
+        await prisma.account.delete({ where: { id: req.body.account } });
+        logger.info({ webhook: req.body }, `Successful webhook event ${WEBHOOKS.ACCOUNT_DELETED} executed`);
+      }
+      break;
+    }
+    case WEBHOOKS.TRACK_OPEN:
+    case WEBHOOKS.MESSAGE_MISSING:
+    case WEBHOOKS.MESSAGE_DELIVERY_ERROR:
+    case WEBHOOKS.MESSAGE_FAILED:
+    case WEBHOOKS.MESSAGE_BOUNCE: {
       const activity = await prisma.activity.findFirst({ where: { messageId: req.body.data.messageId } });
+      const status: Record<string, ActivityStatus> = {
+        [WEBHOOKS.TRACK_OPEN]: ActivityStatus.OPENED,
+        [WEBHOOKS.MESSAGE_BOUNCE]: ActivityStatus.BOUNCED,
+        [WEBHOOKS.MESSAGE_MISSING]: ActivityStatus.ERROR,
+        [WEBHOOKS.MESSAGE_DELIVERY_ERROR]: ActivityStatus.ERROR,
+        [WEBHOOKS.MESSAGE_FAILED]: ActivityStatus.ERROR,
+      };
       if (activity) {
-        payload = await prisma.activity.create({
-          data: { ...omit(activity, ['id', 'createdAt', 'updatedAt', 'queueId']), status: ActivityStatus.OPENED },
+        await prisma.activity.create({
+          data: {
+            ...omit(activity, ['id', 'createdAt', 'updatedAt', 'queueId']),
+            status: status[req.body.event as string],
+          },
         });
       }
       break;
@@ -44,6 +95,30 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
           if (createdActivity.campaign.sendOnReply) {
             //find all queueId for this lead and remove them
           }
+        }
+      }
+      break;
+    }
+    case WEBHOOKS.MESSAGE_SENT: {
+      const activity = await prisma.activity.findFirst({
+        where: { messageId: req.body.data.messageId },
+        include: { campaign: true },
+      });
+      const sequences = activity?.campaign.sequences as Prisma.JsonArray;
+      if (activity) {
+        await prisma.activity.create({
+          data: {
+            ...omit(activity, ['id', 'createdAt', 'updatedAt', 'queueId', 'campaign']),
+            status: ActivityStatus.CONTACTED,
+          },
+        });
+        if (activity.step === sequences.length) {
+          await prisma.activity.create({
+            data: {
+              ...omit(activity, ['id', 'createdAt', 'updatedAt', 'queueId', 'campaign']),
+              status: ActivityStatus.COMPLETED,
+            },
+          });
         }
       }
       break;
